Adjust subscription tests to run with both protocols
