+++ server/integration-tests/src/test/java/io/smallrye/graphql/tests/client/dynamic/subscription/AbstractDynamicClientSubscriptionTest.java
@@ -15,28 +15,21 @@ import java.util.stream.Collectors;
 import javax.json.JsonValue;
 
 import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.asset.EmptyAsset;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.After;
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import io.smallrye.graphql.client.GraphQLError;
 import io.smallrye.graphql.client.Response;
 import io.smallrye.graphql.client.core.Document;
 import io.smallrye.graphql.client.core.OperationType;
 import io.smallrye.graphql.client.vertx.dynamic.VertxDynamicGraphQLClient;
-import io.smallrye.graphql.client.vertx.dynamic.VertxDynamicGraphQLClientBuilder;
 
-@RunWith(Arquillian.class)
-@RunAsClient
-public class DynamicClientSubscriptionTest {
+public abstract class AbstractDynamicClientSubscriptionTest {
 
     @Deployment
     public static WebArchive deployment() {
@@ -48,14 +41,7 @@ public class DynamicClientSubscriptionTest {
     @ArquillianResource
     URL testingURL;
 
-    private static VertxDynamicGraphQLClient client;
-
-    @Before
-    public void prepare() {
-        client = (VertxDynamicGraphQLClient) new VertxDynamicGraphQLClientBuilder()
-                .url(testingURL.toString() + "graphql")
-                .build();
-    }
+    protected static VertxDynamicGraphQLClient client;
 
     @After
     public void cleanup() {
